Add post composition page, update sidebar routes, and enhance post data handling
