+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,a={};const s=(s,t)=>(s=new URL(s+".js",t).href,a[s]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=a,document.head.appendChild(e)}else e=s,importScripts(s),a()})).then((()=>{let e=a[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e})));self.define=(t,c)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(a[n])return;let i={};const b=e=>s(e,n),r={module:{uri:n},exports:i,require:b};a[n]=Promise.all(t.map((e=>r[e]||b(e)))).then((e=>(c(...e),i)))}}define(["./workbox-4754cb34"],(function(e){"use strict";importScripts("worker-wfbTD3R-YMmcSGUe7txab.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/300x300.png",revision:"1b9c0d317b1168666ee7a8f5785ad9d1"},{url:"/_next/app-build-manifest.json",revision:"c545509a7591bbd65e553dde60df6d1a"},{url:"/_next/dynamic-css-manifest.json",revision:"d751713988987e9331980363e24189ce"},{url:"/_next/static/chunks/0e5ce63c-60d54ebb17d9d80d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/13b76428-1947e09b2321d766.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/1505-00997e18565d09b4.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/1684-8eeec9cb80141d4b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/187-50fb081dcc753bf3.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/2154-73011a8776d8daa7.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/2600-baa6b4b25490a961.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/3063-d61f142f1785f03f.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/4260-40542224337d1688.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/4298-b8dd40e4c82efd6b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/495-f67d30e835a26606.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/4bd1b696-ac71ba708f7440fa.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/5275-c19f4ffc2db1eb54.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/5615-8354718a0cd53fd7.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/5709-5db8d3e7933de6cb.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/6421-3aebfc76a9a6f171.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/6443-29cb747e83bde772.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/6505-587f890d70233ec0.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/6874-e80468eb4933642c.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/6914-4a87457d94bc4409.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/7223-36f16cbe6d5bb8ec.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/7408-64b733cfdbf4be23.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/7998-3d85f96d87301ab1.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/870fdd6f-76aecbd427fa26b9.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/9059-d75bc7ea6589500f.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/9669-bf5619d7c0e37fb5.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/aaea2bcf-f58e4c5afad46e25.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/%5Busername%5D/page-d8ff839500a1b37f.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/%5Busername%5D/photo/page-106c1aad9432a27b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/%5Busername%5D/photos/route-1ebbd394958cbf91.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/%5Busername%5D/posts/%5Bid%5D/page-0062ff6e6b0d38bb.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/%5Busername%5D/posts/%5Bid%5D/photo/%5Bindex%5D/page-69822a6cdf541ca0.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/@modal/(.)%5Busername%5D/photo/page-69d27bae0adf1a0b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/@modal/(.)%5Busername%5D/posts/%5Bid%5D/photo/%5Bindex%5D/page-c0964b15d1d8208a.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/@modal/(.)call/page-c5e68e93670c3e03.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/@modal/default-0b339201e5c2cd6a.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/_not-found/page-f16550bdef2e3247.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/account/personal-info/page-0346d7150d3ec26d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/account/privacy-security/page-df2679ed684822d4.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/confirm-email/%5Btoken%5D/route-cef9ca34bb2b1f82.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/confirm-email/failed/page-92ad6932c26e781d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/confirm-email/success/page-39db2e9b57892fa6.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/forgot-password/page-74749df5a033cf1a.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/layout-51d21d680082c339.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/login/page-1ec08ca1f4fb3d67.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/logout/page-63d5e72d929a0e3a.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/reset-password/page-544b501f30f158d9.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/signup-complete/page-e0ad08a4bcaf9906.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/signup/layout-1765760f22e50b95.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/accounts/signup/page-c70d8e13780d1ef8.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-06e77ae8618bb67f.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/getuser/route-b9ea095c89fbbc0d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/link-preview/route-e901ca13db8e9acc.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/logout/route-c2c5954f34de52e9.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/protected/route-e14e447769ffbd29.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/signup/route-ba2f4585639ca8bf.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/api/upload/route-c2374f80976c2c5f.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/auth/error/page-0b001a8d257ff3da.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/call/page-be6ad298e6da20fa.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/layout-9cf85892d08799b3.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/page-5f750ed79a2e56f3.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/settings/page-4d519720f339a26c.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/compose/group/page-c32ee0f3e89b05dc.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/compose/page-fbcadfd70418e33b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/layout-42fbebc064cb919a.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/page-d68178f48941ed32.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/settings/page-55cafd42765bfb53.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/layout-fbcfdc524480c95e.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/chats/page-67899519823c136e.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/default-c7d5baf9f951e426.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/error-a84a356a17b707b2.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/explore/page-18d0475f9b95861e.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/explore/q/page-55ecef8978a7f3e8.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/feedback/page-6d125712f69fb8e0.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/general/page-35162f63a3040abd.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/global-error-3474e26de69abd4d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/help-support/page-bde4802a03b3bc9b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/home/layout-984bedabb9685f79.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/home/page-caed81db817d16cd.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/layout-b06d5611976b3082.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/loading-0e345dd72209bbc0.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/manifest.webmanifest/route-139f61c4e3a4467e.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/not-found-de7ac3e6c6f66c8d.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/notifications/page-ab865710b24d444e.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/page-a91926cd0bc82e39.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/postgres/page-2ff5947851b1721b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/pushNotification/page-136b942eb5ae7de3.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/robots.txt/route-8e1a328f3d9361c2.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/app/sitemap.xml/route-290f8c288d967ea7.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/fcfb803e-64c523f883d3823b.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/framework-033a16219e131880.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/main-app-c4acc63d5ff3bd52.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/main-e68a943c10f5a0fa.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/pages/_app-5d1abe03d322390c.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/pages/_error-3b2a1d523de49635.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-a47fb8f484a11a02.js",revision:"wfbTD3R-YMmcSGUe7txab"},{url:"/_next/static/css/782b6d20e501c7a8.css",revision:"782b6d20e501c7a8"},{url:"/_next/static/css/b7715e5b837d80ba.css",revision:"b7715e5b837d80ba"},{url:"/_next/static/css/bd4fb75cdc5ec81b.css",revision:"bd4fb75cdc5ec81b"},{url:"/_next/static/css/d94a56b6a4bfdc9e.css",revision:"d94a56b6a4bfdc9e"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/6d93bde91c0c2823-s.woff2",revision:"621a07228c8ccbfd647918f1021b4868"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/a34f9d1faa5f3315-s.p.woff2",revision:"d4fe31e6a2aebc06b8d6e558c9141119"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/eb5d8667c55dd7e2-s.p.ttf",revision:"6310192cd2011f527e18b1586a1245c8"},{url:"/_next/static/wfbTD3R-YMmcSGUe7txab/_buildManifest.js",revision:"c7c187e56f640b4d4fdbb11a4aa9e5a0"},{url:"/_next/static/wfbTD3R-YMmcSGUe7txab/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/badge.png",revision:"564a99cc4995491cd775c26e79f6ea28"},{url:"/default.jpeg",revision:"caa54b988b34be28a36aec2f1207829e"},{url:"/emails.hbs",revision:"6172c003c1990e8bae515b98413e3d79"},{url:"/fonts/Raleway-LightItalic.ttf",revision:"f3c1507fc3d8231fa7011d827acd214b"},{url:"/fonts/Raleway-Medium.ttf",revision:"28ea37f0eb58c57e01eed0b06fc359d6"},{url:"/fonts/Raleway-Regular.ttf",revision:"6310192cd2011f527e18b1586a1245c8"},{url:"/fonts/Raleway-SemiBold.ttf",revision:"66c9748f1e4aae2e764d5c50c05f7841"},{url:"/greybg-dark.png",revision:"90de21142f126cfd1ecb25332e8c701d"},{url:"/greybg.jpg",revision:"a0b538e2be702d8fcafc5321a43c6311"},{url:"/heart.png",revision:"9f61111962cc66e53f0c907d4fcc17cb"},{url:"/home.png",revision:"ba12db4bf951a7e43ce4918554d81db5"},{url:"/largeScreenShot_home.png",revision:"282740f4f0a17b4181dd355117727ece"},{url:"/messages.png",revision:"e4d5b2c4d733b0cf27a4071c11ad9c5e"},{url:"/mobileScreenShot_home.png",revision:"0a3e22ff06f42e252f942b5a19e01f8b"},{url:"/notifications.png",revision:"ce4468ad7db94d60ee00ad02d4c5c598"},{url:"/profile.png",revision:"4907111c2900c470379dfa35e9503f53"},{url:"/settings.png",revision:"7ba0d400446a18918ceea0cbdfaaa136"},{url:"/velo10.png",revision:"f4d3f4f9a5d7df497ebe8a0227e0b62f"},{url:"/velo11.png",revision:"ec7ef09a039dddba1a2cf2d46810c5ba"},{url:"/velo9.png",revision:"421fba9427d9671c9cac4f16f0209cbb"},{url:"/verified.svg",revision:"99678523a3b1317a387b8d8f2951fb08"},{url:"/web-app-manifest-144x144.png",revision:"28380770b36ef80c667c28b976328ba7"},{url:"/web-app-manifest-192x192.png",revision:"18f1c98ce197c78e0a0979aad24c55b7"},{url:"/web-app-manifest-512x512.png",revision:"09b14f9185fc5516fa1913addd4c9645"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:s,state:t})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,t)=>(a=new URL(a+".js",t).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(t,n)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let c={};const r=e=>a(e,i),u={module:{uri:i},exports:c,require:r};s[i]=Promise.all(t.map((e=>u[e]||r(e)))).then((e=>(n(...e),c)))}}define(["./workbox-4754cb34"],(function(e){"use strict";importScripts("worker-Y6Pj3e2B6T4SvaPHNwPet.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/300x300.png",revision:"1b9c0d317b1168666ee7a8f5785ad9d1"},{url:"/_next/app-build-manifest.json",revision:"99675df64c935026423f6d445151cf17"},{url:"/_next/dynamic-css-manifest.json",revision:"d751713988987e9331980363e24189ce"},{url:"/_next/static/Y6Pj3e2B6T4SvaPHNwPet/_buildManifest.js",revision:"6f9d82ed4ef8bcf2d7e43a8d7bc20dbf"},{url:"/_next/static/Y6Pj3e2B6T4SvaPHNwPet/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/0e5ce63c-60d54ebb17d9d80d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/13b76428-1947e09b2321d766.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/1505-00997e18565d09b4.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/1684-8eeec9cb80141d4b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/187-50fb081dcc753bf3.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/1970-af2359cdc3f43548.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/2397-723b8397aaa2784e.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/2600-baa6b4b25490a961.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/3063-d61f142f1785f03f.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/4260-40542224337d1688.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/4298-b8dd40e4c82efd6b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/495-f67d30e835a26606.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/4bd1b696-ac71ba708f7440fa.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/5275-b10e1b054af3c054.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/5615-8354718a0cd53fd7.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/5709-5db8d3e7933de6cb.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6337-4fa50816a5f16add.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6421-3aebfc76a9a6f171.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6443-29cb747e83bde772.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6505-587f890d70233ec0.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6874-e80468eb4933642c.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/6914-4a87457d94bc4409.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/7223-36f16cbe6d5bb8ec.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/7998-3d85f96d87301ab1.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/870fdd6f-76aecbd427fa26b9.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/9059-d75bc7ea6589500f.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/9669-bf5619d7c0e37fb5.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/aaea2bcf-f58e4c5afad46e25.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/%5Busername%5D/page-9d090d90ca4626ef.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/%5Busername%5D/photo/page-276e4dbe2a90da80.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/%5Busername%5D/photos/route-f7ef28a05be8f4d5.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/%5Busername%5D/posts/%5Bid%5D/page-767418999856e6a5.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/%5Busername%5D/posts/%5Bid%5D/photo/%5Bindex%5D/page-923bbee5c312fc6c.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/@modal/(.)%5Busername%5D/photo/page-2c32ca65b8536e57.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/@modal/(.)%5Busername%5D/posts/%5Bid%5D/photo/%5Bindex%5D/page-ec60d668a0107897.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/@modal/(.)call/page-ee1f3ed050398dd2.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/@modal/(.)compose/post/page-23705ab145563e0b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/@modal/default-bb44607960ddb413.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/_not-found/page-4f463749660202b9.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/account/personal-info/page-0346d7150d3ec26d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/account/privacy-security/page-df2679ed684822d4.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/confirm-email/%5Btoken%5D/route-a774be4e54336c47.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/confirm-email/failed/page-92ad6932c26e781d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/confirm-email/success/page-39db2e9b57892fa6.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/forgot-password/page-74749df5a033cf1a.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/layout-51d21d680082c339.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/login/page-1ec08ca1f4fb3d67.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/logout/page-63d5e72d929a0e3a.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/reset-password/page-544b501f30f158d9.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/signup-complete/page-e0ad08a4bcaf9906.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/signup/layout-5d69f0d48f35abd3.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/accounts/signup/page-c70d8e13780d1ef8.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-8fd0d0dae85ebc15.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/getuser/route-337acb31192d7929.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/link-preview/route-3f871df5e79a23f2.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/logout/route-67e7484100e7ed37.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/protected/route-713e04d881bcfadd.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/signup/route-cbceb39f30b1423d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/api/upload/route-435a12d8ca4237f0.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/auth/error/page-0b001a8d257ff3da.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/call/page-54d53c343c37bf3b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/layout-34eaa6c4633bf8a8.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/page-5f750ed79a2e56f3.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/%5Bid%5D/settings/page-4d519720f339a26c.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/compose/group/page-fc54ff79167274d6.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/compose/page-c37cb79d871e0056.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/layout-823804eb2e88e61e.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/page-7e743feee585601c.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/group/%5Bgid%5D/settings/page-55cafd42765bfb53.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/layout-42af30e2317250a1.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/chats/page-67899519823c136e.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/compose/post/page-de1f2fd9fae2e7a3.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/default-37504a9bd8efbab2.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/error-a84a356a17b707b2.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/explore/page-18d0475f9b95861e.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/explore/q/page-55ecef8978a7f3e8.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/feedback/page-6d125712f69fb8e0.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/general/page-35162f63a3040abd.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/global-error-3474e26de69abd4d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/help-support/page-bde4802a03b3bc9b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/home/layout-c5815b772db15d0b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/home/page-ad9539e0a655596d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/layout-6fc4dabf3967a181.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/loading-0e345dd72209bbc0.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/manifest.webmanifest/route-026c0be0b39c04dd.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/not-found-de7ac3e6c6f66c8d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/notifications/page-ab865710b24d444e.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/page-a91926cd0bc82e39.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/postgres/page-5d533f9aa3b674e2.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/pushNotification/page-136b942eb5ae7de3.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/robots.txt/route-c6a05eb3f7ee0f3d.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/app/sitemap.xml/route-07912a270be23914.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/fcfb803e-64c523f883d3823b.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/framework-033a16219e131880.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/main-app-c4acc63d5ff3bd52.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/main-e68a943c10f5a0fa.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/pages/_app-5d1abe03d322390c.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/pages/_error-3b2a1d523de49635.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-e37e1dc9454bf540.js",revision:"Y6Pj3e2B6T4SvaPHNwPet"},{url:"/_next/static/css/02d7dcc813861f60.css",revision:"02d7dcc813861f60"},{url:"/_next/static/css/b7715e5b837d80ba.css",revision:"b7715e5b837d80ba"},{url:"/_next/static/css/bd4fb75cdc5ec81b.css",revision:"bd4fb75cdc5ec81b"},{url:"/_next/static/css/d94a56b6a4bfdc9e.css",revision:"d94a56b6a4bfdc9e"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/6d93bde91c0c2823-s.woff2",revision:"621a07228c8ccbfd647918f1021b4868"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/a34f9d1faa5f3315-s.p.woff2",revision:"d4fe31e6a2aebc06b8d6e558c9141119"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/eb5d8667c55dd7e2-s.p.ttf",revision:"6310192cd2011f527e18b1586a1245c8"},{url:"/badge.png",revision:"564a99cc4995491cd775c26e79f6ea28"},{url:"/default.jpeg",revision:"caa54b988b34be28a36aec2f1207829e"},{url:"/emails.hbs",revision:"6172c003c1990e8bae515b98413e3d79"},{url:"/fonts/Raleway-LightItalic.ttf",revision:"f3c1507fc3d8231fa7011d827acd214b"},{url:"/fonts/Raleway-Medium.ttf",revision:"28ea37f0eb58c57e01eed0b06fc359d6"},{url:"/fonts/Raleway-Regular.ttf",revision:"6310192cd2011f527e18b1586a1245c8"},{url:"/fonts/Raleway-SemiBold.ttf",revision:"66c9748f1e4aae2e764d5c50c05f7841"},{url:"/greybg-dark.png",revision:"90de21142f126cfd1ecb25332e8c701d"},{url:"/greybg.jpg",revision:"a0b538e2be702d8fcafc5321a43c6311"},{url:"/heart.png",revision:"9f61111962cc66e53f0c907d4fcc17cb"},{url:"/home.png",revision:"ba12db4bf951a7e43ce4918554d81db5"},{url:"/largeScreenShot_home.png",revision:"282740f4f0a17b4181dd355117727ece"},{url:"/messages.png",revision:"e4d5b2c4d733b0cf27a4071c11ad9c5e"},{url:"/mobileScreenShot_home.png",revision:"0a3e22ff06f42e252f942b5a19e01f8b"},{url:"/notifications.png",revision:"ce4468ad7db94d60ee00ad02d4c5c598"},{url:"/profile.png",revision:"4907111c2900c470379dfa35e9503f53"},{url:"/settings.png",revision:"7ba0d400446a18918ceea0cbdfaaa136"},{url:"/velo10.png",revision:"f4d3f4f9a5d7df497ebe8a0227e0b62f"},{url:"/velo11.png",revision:"ec7ef09a039dddba1a2cf2d46810c5ba"},{url:"/velo9.png",revision:"421fba9427d9671c9cac4f16f0209cbb"},{url:"/verified.svg",revision:"99678523a3b1317a387b8d8f2951fb08"},{url:"/web-app-manifest-144x144.png",revision:"28380770b36ef80c667c28b976328ba7"},{url:"/web-app-manifest-192x192.png",revision:"18f1c98ce197c78e0a0979aad24c55b7"},{url:"/web-app-manifest-512x512.png",revision:"09b14f9185fc5516fa1913addd4c9645"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:t})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
