+++ src/app/@modal/(.)compose/post/page.tsx
@@ -1,0 +1,3 @@
+export default function post () {
+    return null;
+}
