+++ src/components/ShareButton.tsx
@@ -19,6 +19,7 @@ import { Button } from "./ui/button";
 import { DrawerTrigger, DrawerContent, DrawerHeader, DrawerTitle, DrawerDescription, DrawerFooter, DrawerClose, Drawer } from "./ui/drawer";
 import { Skeleton } from "./ui/skeleton";
 import { Input } from "./ui/input";
+import PostMaker from "./PostMaker";
 
 interface Option {
   icon: React.ReactNode;
@@ -126,149 +127,13 @@ export default function ShareButton({children, post} : {children: ReactNode, pos
       setIsPopoverOpen(open);
     };
   
-    function handleFiles(e: ChangeEvent<HTMLInputElement>): void {
-      if (!imageInputRef.current) return;
-  
-      const mediaFiles = e.target.files;
-      if (!mediaFiles) return;
-  
-      const newFiles = Array.from(mediaFiles);
-      const validationErrors: string[] = [];
-      const validFiles: File[] = [];
-  
-      // Check total number of files
-      if (files.length > FILE_VALIDATION_CONFIG.maxFiles) {
-        toast({
-          title: 'Warning',
-          description: `Maximum ${FILE_VALIDATION_CONFIG.maxFiles} files allowed`,
-          variant: 'destructive'
-        });
-        clearFiles();
-        return;
-      }
-  
-      // Calculate total size including existing files
-      const existingSize = files.reduce((acc, file) => acc + file.size, 0);
-      const newTotalSize = newFiles.reduce((acc, file) => acc + file.size, existingSize);
-  
-      if (newTotalSize > FILE_VALIDATION_CONFIG.maxTotalSize) {
-        toast({
-          title: 'Warning',
-          description: `Total size exceeds ${formatFileSize(FILE_VALIDATION_CONFIG.maxTotalSize)}`,
-          variant: 'destructive'
-        });
-        return;
-      }
-  
-      // Validate each file
-      newFiles.forEach(file => {
-        try {
-          validateFile(file, FILE_VALIDATION_CONFIG);
-          validFiles.push(file);
-        } catch (error) {
-          if (error instanceof Error) {
-            validationErrors.push(error.message);
-          }
-        }
-      });
-  
-      // Update state
-      if (validationErrors.length > 0) {
-        setErrors(validationErrors);
-        validationErrors.forEach(error => toast({
-          title: 'Warning',
-          description: error
-        }));
-      }
-  
-      if (validFiles.length > 0) {
-        setFiles(validFiles);
-      }
-  
-      // Reset input
-      e.target.value = '';
-    }
-  
-  
-    const handlePost = async () => {
-      if(!socket || !post) return;
-      setIsPosting(true)
-      
-      try {
-        let media: string[] = []
-        if(files.length > 0) {
-  
-          files.forEach(async (file) => {
-            const response = await fetch(
-              '/api/upload',
-              {
-                method: 'POST',
-                headers: {
-                  'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({ filename: file!.name, contentType: file!.type, bucketName: 'post-s' }),
-              }
-            );
-    
-            if (!response.ok) {
-              throw new Error('Failed to upload image');
-            }
-    
-            const { url, fields } = await response.json();
-            const formData1 = new FormData();
-    
-            for (const key in fields) {
-              formData1.append(key, fields[key]);
-            }
-    
-            formData1.append('file', file);
-    
-            const uploadResponse = await fetch(url, {
-              method: 'POST',
-              body: formData1,
-            });
-    
-            if (!uploadResponse.ok) { 
-              throw new Error('Failed to upload image');
-            } else {
-              // formData.file = url+fields.key;
-              media.push(url+fields.key);
-            }
-          })
-  
-        }
-  
-        const quote: Partial<PostData> = {
-          Visibility: visibility as PostData['Visibility'],
-          Caption: text,
-          Image: media,
-          Code: '',
-          WhoCanComment: visibility as PostData['WhoCanComment'],
-          OriginalPostId: post.PostID
-        }
-  
-        socket.emit('reactToPost(share)', {
-          action: "share",
-          type: "quote",
-          post: quote
-        });
-  
-      } catch (error: any) {
-        console.log(error)
-      } finally {
-        setFiles([]);
-        setIsQuoteModalOpen(false);
-        setIsPosting(false)
-      }
-    };
-  
     if (!userdata._id) return <div className='flex items-center justify-center gap-2 cursor-not-allowed'>{children}</div>;
   
     return (
       <>
         {/* Mobile view */}
         <Drawer open={isDrawerOpen} onOpenChange={handleDrawerChange}>
-          <DrawerTrigger className='tablets:hidden flex items-center justify-center gap-2 cursor-pointer'>
+          <DrawerTrigger className='tablets:hidden blog-foot flex items-center justify-center gap-2 cursor-pointer'>
             {children}
           </DrawerTrigger>
           <DrawerContent className='tablets:hidden'>
@@ -276,19 +141,6 @@ export default function ShareButton({children, post} : {children: ReactNode, pos
               <DrawerTitle className='hidden'>Options</DrawerTitle>
               <DrawerDescription className='flex flex-col gap-2'>
                 {options.map(({ icon, text, onClick }, index) => {
-                  if(text === "Quote") return (
-                    <span 
-                      key={index} 
-                      className='flex gap-1 items-center cursor-pointer w-full' 
-                      onClick={() => {
-                        setIsDrawerOpen(false);
-                        setIsQuoteModalOpen(true);
-                      }}
-                    >
-                      {icon}
-                      <span className='dark:text-white'>{text}</span>
-                    </span>
-                  )
                   return (
                     <span key={index} className='flex gap-1 items-center cursor-pointer w-full' onClick={onClick}>
                       {icon}
@@ -308,24 +160,11 @@ export default function ShareButton({children, post} : {children: ReactNode, pos
   
         {/* Desktop view */}
         <DropdownMenu open={isPopoverOpen} onOpenChange={handlePopoverChange}>
-          <DropdownMenuTrigger className='hidden tablets:flex items-center justify-center gap-2 cursor-pointer'>
+          <DropdownMenuTrigger className='hidden blog-foot tablets:flex items-center justify-center gap-2 cursor-pointer'>
             {children}
           </DropdownMenuTrigger>
           <DropdownMenuContent className='bg-white hidden tablets:block dark:bg-zinc-800 w-auto space-y-2 mt-1 mr-2 p-2 rounded-md shadow-lg'>
             {options.map(({ icon, text, onClick }, index) => {
-              if(text === "Quote") return (
-                <Button 
-                  key={index} 
-                  className='flex gap-1 items-center cursor-pointer w-full bg-white dark:bg-zinc-900 text-black dark:text-white hover:bg-white/90 dark:hover:bg-zinc-900/90' 
-                  onClick={() => {
-                    setIsPopoverOpen(false);
-                    setIsQuoteModalOpen(true);
-                  }}
-                >
-                  {icon}
-                  <span className='dark:text-white'>{text}</span>
-                </Button>
-              )
               return (
                 <Button key={index} className='flex gap-1 items-center cursor-pointer w-full bg-white dark:bg-zinc-900 text-black dark:text-white hover:bg-white/90 dark:hover:bg-zinc-900/90' onClick={onClick}>
                   {icon}
@@ -337,210 +176,7 @@ export default function ShareButton({children, post} : {children: ReactNode, pos
         </DropdownMenu>
   
         {/* Separate Dialog component */}
-        <Dialog open={isQuoteModalOpen} onOpenChange={setIsQuoteModalOpen}>
-          <DialogContent className="bg-white dark:bg-zinc-900 overflow-auto mb:h-full">
-            
-            <DialogHeader className='dark:text-white'>
-              <DialogTitle className='text-center'></DialogTitle>
-              <DialogDescription>
-                <span className="flex justify-between items-center">
-                  <DialogClose className="dark:text-white hover:bg-gray-800 p-2 rounded-full transition-all duration-200 transform hover:scale-110">
-                    <X size={16} />
-                  </DialogClose>
-                  <span className="text-brand text-xl font-bold">Drafts</span>
-                </span>
-              </DialogDescription>
-            </DialogHeader>
-            <div className="max-h-80 overflow-auto">
-  
-              {/* User Info and Dropdown */}
-              <div className="flex items-center space-x-3">
-                <div className="w-10 h-10 rounded-full overflow-hidden">
-                  {userdata.dp ?
-                  <img
-                    src={userdata.dp}
-                    alt="Profile"
-                    className="w-full h-full object-cover"
-                  /> :
-                  <Skeleton className="w-full h-full object-cover"/>}
-                </div>
-                <Select onValueChange={setVisibility} defaultValue="everyone" value={visibility}>
-                  <SelectTrigger className="max-w-[120px] rounded-2xl">
-                    <SelectValue placeholder="Everyone" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="everyone">Everyone</SelectItem>
-                    <SelectItem value="none">None</SelectItem>
-                    <SelectItem value="friends">Friends</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-  
-              <Input
-                accept=".jpg, .jpeg, .png, .gif, .mp4, .mov, .webm, .mp3, .wav, .ogg"
-                id="mediaInput"
-                ref={imageInputRef} 
-                type="file"
-                multiple 
-                onChange={handleFiles} 
-                className="hidden"
-              />
-  
-              {/* Comment Input */}
-              <textarea
-                ref={textAreaRef} 
-                value={text}
-                maxLength={textLimit}
-                spellCheck
-                onChange={(e) => setText(e.target.value.substring(0, textLimit))}
-                placeholder="Add a comment" 
-                className="flex-grow text-sm w-full mt-2 max-h-[400px] focus:border-b focus:border-brand resize-none bg-transparent border-none focus:outline-none dark:text-white"
-                onInput={handleInput}
-              />
-
-              {/* Fullscreen Image Modal */}
-                {fullscreenImage && (
-                    <Dialog open={!!fullscreenImage} onOpenChange={closeFullscreen}>
-                    <DialogContent className="bg-black flex items-center justify-center">
-                        <DialogClose className="absolute top-4 right-4 text-white">
-                        <X size={24} />
-                        </DialogClose>
-                        <img
-                        src={fullscreenImage}
-                        alt="Fullscreen"
-                        className="max-w-full max-h-full object-contain"
-                        />
-                    </DialogContent>
-                    </Dialog>
-                )}
-  
-              {/** Media */}
-              {files.length > 0 && (
-                <div className="grid grid-cols-2 gap-2 my-2">
-                  {files.map((file, index) => (
-                    <div key={index} className="relative group">
-                      {file.type.startsWith('image/') ? (
-                        <>
-                            <img
-                            src={URL.createObjectURL(file)}
-                            alt={file.name}
-                            className="w-full h-40 object-cover rounded-lg"
-                            onClick={() => handleImageClick(file)}
-                            />
-                            <CropMediaInterface files={files} setFiles={setFiles} imageIndex={index}>
-                                <button
-                                onClick={() => {
-                                    setIsEditOpen(true)
-                                }}
-                                className="absolute top-2 left-2 p-1 rounded-full bg-black/50 text-white opacity-0 group-hover:opacity-100 transition-opacity"
-                                >
-                                <Paintbrush size={16} />
-                                </button>
-                            </CropMediaInterface>
-                        </>
-                      ) : file.type.startsWith('video/') ? (
-                        <video
-                          src={URL.createObjectURL(file)}
-                          className="w-full h-40 object-cover rounded-lg"
-                          controls
-                        />
-                      ) : null}
-                      <button
-                        onClick={() => {
-                          const newFiles = files.filter((_, i) => i !== index);
-                          setFiles(newFiles);
-                        }}
-                        className="absolute top-2 right-2 p-1 rounded-full bg-black/50 text-white opacity-0 group-hover:opacity-100 transition-opacity"
-                      >
-                        <X size={16} />
-                      </button>
-                    </div>
-                  ))}
-                </div>
-              )}
-  
-              {/* Post Card */}
-              {post && (
-                <div className="border border-gray-800 rounded-xl p-4">
-                  <div className="flex items-start space-x-3">
-                    <div className="w-10 h-10 rounded-full overflow-hidden flex-shrink-0">
-                      <img
-                        src={post.DisplayPicture}
-                        alt={post.Username}
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <div className="flex-1 grid gap-2 grid-cols-2">
-                      <div className="flex items-center col-span-2 flex-wrap">
-                        <span className="font-bold dark:text-white mr-1 truncate">{post.NameOfPoster}</span>
-                        {post.Verified && (
-                          <svg className="w-4 h-4 text-brand fill-current" viewBox="0 0 24 24">
-                            <path d="M22.5 12.5c0-1.58-.875-2.95-2.148-3.6.154-.435.238-.905.238-1.4 0-2.21-1.71-3.998-3.818-3.998-.47 0-.92.084-1.336.25C14.818 2.415 13.51 1.5 12 1.5s-2.816.917-3.437 2.25c-.415-.165-.866-.25-1.336-.25-2.11 0-3.818 1.79-3.818 4 0 .494.083.964.237 1.4-1.272.65-2.147 2.018-2.147 3.6 0 1.495.782 2.798 1.942 3.486-.02.17-.032.34-.032.514 0 2.21 1.708 4 3.818 4 .47 0 .92-.086 1.335-.25.62 1.334 1.926 2.25 3.437 2.25 1.512 0 2.818-.916 3.437-2.25.415.163.865.248 1.336.248 2.11 0 3.818-1.79 3.818-4 0-.174-.012-.344-.033-.513 1.158-.687 1.943-1.99 1.943-3.484zm-6.616-3.334l-4.334 6.5c-.145.217-.382.334-.625.334-.143 0-.288-.04-.416-.126l-.115-.094-2.415-2.415c-.293-.293-.293-.768 0-1.06s.768-.294 1.06 0l1.77 1.767 3.825-5.74c.23-.345.696-.436 1.04-.207.346.23.44.696.21 1.04z" />
-                        </svg>
-                        )}
-                        <span className="text-gray-500 ml-1">@{post.Username} · {time}</span>
-                      </div>
-                      {/* <div className="text-gray-500 text-sm mb-2">Replying to @NintendoAmerica</div> */}
-                      {post.Caption ? (
-                        <p className={`dark:text-white text-sm mb-2 ${post.Image.length > 0 ? '' : 'col-span-2'} whitespace-pre-wrap`}>{post.Caption.length > 250 ? post.Caption.substring(0, 250) + '...' : post.Caption}</p>
-                      ) : null}
-                      {/* {showMore} */}
-                      {(post.Image.length > 0) &&
-                        <MediaSlide className={`overflow-auto rounded-lg ${!post.Caption.length ? 'col-span-2' : ''}`} postData={post}/>
-                      }
-                    </div>
-                  </div>
-                </div>
-              )}
-            </div>
-            
-            <div className='sticky bottom-0 space-y-2'>
-              {/* Everyone Can Reply */}
-              <div className="bg-white dark:bg-zinc-900 flex items-center text-brand space-x-2">
-                <CircleAlert size={16} />
-                <span>{visibility[0].toUpperCase() + visibility.slice(1)} can reply</span>
-              </div>
-  
-              {/* Character Counter */}
-              <div className={`flex justify-end text-gray-500 text-sm mb-2 ${text.length === textLimit ? 'text-red-500' : ''}`}>
-                <span>{text.length === textLimit && 'You have reached the text characters limit! • '}{text.length}/{textLimit}</span>
-              </div>
-  
-              <DialogFooter className='bg-white dark:bg-zinc-900'>
-                {/* Bottom Toolbar */}
-                <div className="border-t border-gray-800 flex-1 py-3">
-                  <div className="gap-8 grid grid-cols-2 mb:grid-cols-1">
-                    <div className="flex justify-between items-center">
-                      <button onClick={() => imageInputRef?.current?.click()} className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <Images size={16} className='group-hover:text-white'/>
-                      </button>
-                      <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <CircleCheck size={16} className='group-hover:text-white'/>
-                      </button>
-                      <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <ChartBarDecreasing size={16} className='group-hover:text-white'/>
-                      </button>
-                      <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <Smile size={16} className='group-hover:text-white'/>
-                      </button>
-                      <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <Clock4 size={16} className='group-hover:text-white'/>
-                      </button>
-                      <button className="text-brand hover:bg-brand/90 group p-2 rounded-full transition-all duration-200">
-                        <MapPin size={16} className='group-hover:text-white'/>
-                      </button>
-                    </div>
-  
-                    <Button disabled={!txtButton || text.length === textLimit || !text.length} onClick={handlePost} className="bg-brand hover:bg-brand/60 text-white w-full my-2 tablets:w-auto tablets:my-0 font-bold py-2 px-6 rounded-full transition-all duration-200 transform hover:scale-105">
-                      {isPosting ? 'Posting...' : 'Post'}
-                      {isPosting && <Loader2 className='animate-spin ml-2' size={16} />}
-                    </Button>
-                  </div>
-                </div>
-              </DialogFooter>
-            </div>
-          </DialogContent>
-        </Dialog> 
+            <PostMaker type="quote" post={post} open={isQuoteModalOpen} onOpenChange={setIsQuoteModalOpen}/>
       </>
     )
   }
