+++ src/lib/apiUtils.ts
@@ -17,7 +17,12 @@ export async function addInteractionFlags(db: Db, posts: WithId<Document>[], use
       await Promise.all(
         Object.entries(collectionsMap).map(async ([collectionName, field]) => {
           const collection = db.collection(`Posts_${collectionName}`);
-          const result = await collection.findOne({ postId: post.PostID, userId: userId });
+          const result = await (async () => {
+            if (collectionName === 'Shares') {
+              return collection.findOne({ OriginalPostId: post.PostID, UserId: userId });
+            }
+            return collection.findOne({ postId: post.PostID, userId: userId });
+          })();
 
           if (result) {
             post[field] = true;
