+++ src/pages/api/comments/[postId].ts
@@ -1,24 +1,29 @@
-import { getMongoClient } from '@/lib/mongodb';
+import { addInteractionFlags } from '@/lib/apiUtils';
+import { verifyToken } from '@/lib/auth';
+import { getMongoClient, getMongoDb } from '@/lib/mongodb';
+import { Payload } from '@/lib/types/type';
+import { ObjectId } from 'mongodb';
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse){
     if(req.method !== 'POST') return res.status(405).json({ message: 'Method Not Allowed' });
 
     try {
-        const token = req.query.postId;
-        const pass = decodeURIComponent(req.cookies.user ? req.cookies.user : '').replace(/"/g, '');
+        const postId = req.query.postId;
+        const cookie = decodeURIComponent(req.cookies.velo_12 ? req.cookies.velo_12 : '').replace(/"/g, '');
+        const payload = await verifyToken(cookie as unknown as string) as unknown as Payload;
         // await client.connect();
         console.log('Server is ready for work');
 
         // Get the collection
-        const client = await getMongoClient();
-        const collection = client.db('mydb').collection('Users');
-        const commentsCollection = client.db('mydb').collection('Posts_Comments');
+        const db = await getMongoDb();
+        const collection = db.collection('Users');
+        const commentsCollection = db.collection('Posts_Comments');
 
 
-        // console.log('165', tokenn)
-        const comments = await commentsCollection.find({ ParentId: token }).toArray();
-        const user = await collection.findOne({ loginToken: pass });
+        // console.log('165', postId)
+        const comments = await commentsCollection.find({ ParentId: postId }).toArray();
+        const user = payload ? await collection.findOne({ _id: new ObjectId(payload._id) }) : null;
         if (comments) {
             comments.map(async (comment) => {
                 // console.log(comment)
@@ -27,6 +32,7 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse){
             })
 
             if (user) {
+                await addInteractionFlags(db, comments, user._id.toString())
                 res.status(200).json({ comments, message: "Success" });
             } else {
                 res.status(200).json({ comments, messge: "Sign in" });
