+++ src/pages/api/post/[id].ts
@@ -1,5 +1,7 @@
+import { verifyToken } from '@/lib/auth';
 import { getMongoDb } from '@/lib/mongodb';
-import { Db } from 'mongodb';
+import { Payload } from '@/lib/types/type';
+import { Db, ObjectId } from 'mongodb';
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse){
@@ -7,37 +9,45 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse){
 
       try{
         const cookie = decodeURIComponent(req.cookies.velo_12 ? req.cookies.velo_12 : '').replace(/"/g, '');
+        const payload = await verifyToken(cookie) as unknown as Payload;
+
+        if (!req.query.id) return res.status(400).json({ error: `Post ID is required` });
         const id = req.query.id;
+
         const db = await getMongoDb();
         console.log('Connected to post.app');
         const post = await fetchPostsFromMultipleCollections(db, `${id}`);
 
         if(post){
           const users = db.collection('Users');
-          const user = await users.findOne({ loginToken: cookie });
-            if (cookie && user) {
-              const collectionsMap = {
-                Likes: 'Liked',
-                Shares: 'Shared',
-                Bookmarks: 'Bookmarked'
-              };
+          const user = payload ? await users.findOne({ _id: new ObjectId(payload._id) }) : null;
 
-              await Promise.all(
-                Object.entries(collectionsMap).map(async ([collectionName, field]) => {
-                  const collection = db.collection(`Posts_${collectionName}`);
-                  const result = await (async () => {
-                    return collection.findOne({ postId: post.PostID, userId: user._id.toString() });
-                  })();
-                  if (result) {
-                    post[field] = true;
+          if (user) {
+            const collectionsMap = {
+              Likes: 'Liked',
+              Shares: 'Shared',
+              Bookmarks: 'Bookmarked'
+            };
+
+            await Promise.all(
+              Object.entries(collectionsMap).map(async ([collectionName, field]) => {
+                const collection = db.collection(`Posts_${collectionName}`);
+                const result = await (async () => {
+                  if (collectionName === 'Shares') {
+                    return collection.findOne({ OriginalPostId: post.PostID, UserId: user._id.toString() });
                   }
-                })
-              );
+                  return collection.findOne({ postId: post.PostID, userId: user._id.toString() });
+                })();
+                if (result) {
+                  post[field] = true;
+                }
+              })
+            );
 
-              return res.status(200).json({ post });
-            } else {
-              return res.status(200).json({ post, message: 'Disable Comment.' });
-            }
+            return res.status(200).json({ post });
+          } else {
+            return res.status(200).json({ post, message: 'Disable Comment.' });
+          }
         }
         return res.status(400).json({ error: 'Oops! Post not found. You can search for other things.' });
       }
